requirements.txt
aiohttp==3.9.1
discord.py==2.5.2
emoji==2.12.1
protobuf==6.30.1
python-dotenv==1.0.1
Requests==2.32.3
